ugene-1.17.0

+++ installer/macosx/bundle_common.sh
@@ -4,6 +4,7 @@ function add-plugin {
 
     PLUGIN_LIB="lib${plugin}.dylib"
     PLUGIN_DESC="${plugin}.plugin"
+    PLUGIN_LICENSE="${plugin}.license"
 
     if [ ! -f ${RELEASE_DIR}/plugins/${PLUGIN_LIB} ] ;  
     then  
@@ -16,10 +17,16 @@ function add-plugin {
         echo "Plugin descriptor file not found: ${PLUGIN_DESC} !"
         exit 1
     fi
-    
+
+    if [ ! -f ${RELEASE_DIR}/plugins/${PLUGIN_LICENSE} ] ; 
+    then
+        echo "Plugin descriptor file not found: ${PLUGIN_LICENSE} !"
+        exit 1
+    fi
+
     cp "${RELEASE_DIR}/plugins/${PLUGIN_LIB}"  "${TARGET_EXE_DIR}/plugins/"
     cp "${RELEASE_DIR}/plugins/${PLUGIN_DESC}" "${TARGET_EXE_DIR}/plugins/"
-    changeQtInstallNames "plugins/${PLUGIN_LIB}"
+    cp "${RELEASE_DIR}/plugins/${PLUGIN_LICENSE}" "${TARGET_EXE_DIR}/plugins/"
     changeCoreInstallNames "plugins/${PLUGIN_LIB}"
 }
 
@@ -36,7 +43,6 @@ function add-library {
     fi
 
     cp "${RELEASE_DIR}/${LIB_FILE}"  "${TARGET_EXE_DIR}/"
-    changeQtInstallNames "${LIB_FILE}"
     changeCoreInstallNames "${LIB_FILE}"		
 }
 
@@ -48,16 +54,18 @@ changeCoreInstallNames () {
         echo "Changing core libs install names for $1"
          	       
         install_name_tool -change libU2Algorithm.1.dylib  @executable_path/libU2Algorithm.1.dylib "$TARGET_EXE_DIR"/$1
-       	install_name_tool -change libU2Core.1.dylib  @executable_path/libU2Core.1.dylib "$TARGET_EXE_DIR"/$1
+        install_name_tool -change libU2Core.1.dylib  @executable_path/libU2Core.1.dylib "$TARGET_EXE_DIR"/$1
         install_name_tool -change libU2Designer.1.dylib  @executable_path/libU2Designer.1.dylib "$TARGET_EXE_DIR"/$1
         install_name_tool -change libU2Formats.1.dylib  @executable_path/libU2Formats.1.dylib "$TARGET_EXE_DIR"/$1
         install_name_tool -change libU2Gui.1.dylib  @executable_path/libU2Gui.1.dylib "$TARGET_EXE_DIR"/$1
         install_name_tool -change libU2Lang.1.dylib  @executable_path/libU2Lang.1.dylib "$TARGET_EXE_DIR"/$1
         install_name_tool -change libU2Private.1.dylib  @executable_path/libU2Private.1.dylib "$TARGET_EXE_DIR"/$1
         install_name_tool -change libU2Remote.1.dylib  @executable_path/libU2Remote.1.dylib "$TARGET_EXE_DIR"/$1
+        install_name_tool -change libU2Script.1.dylib  @executable_path/libU2Script.1.dylib "$TARGET_EXE_DIR"/$1
         install_name_tool -change libU2Test.1.dylib  @executable_path/libU2Test.1.dylib "$TARGET_EXE_DIR"/$1
         install_name_tool -change libU2View.1.dylib  @executable_path/libU2View.1.dylib "$TARGET_EXE_DIR"/$1
         install_name_tool -change libugenedb.1.dylib  @executable_path/libugenedb.1.dylib "$TARGET_EXE_DIR"/$1
+        install_name_tool -change libgtest.1.dylib  @executable_path/libgtest.1.dylib "$TARGET_EXE_DIR"/$1
   
    else
        echo "changeCoreInstallNames: no parameter passed."
@@ -67,29 +75,45 @@ changeCoreInstallNames () {
 }
 
 
+# This function replaces @loader_path with @executable_path for UGENE plugins
+# for the specified Qt library
+# Supposed that Qt is build as frameworks
+restorePluginsQtInstallName () {
+    if [ "$1" ] && [ "$2" ]
+    then
+        install_name_tool -change @loader_path/../../Frameworks/$1.framework/Versions/4/$1 @executable_path/../Frameworks/$1.framework/Versions/4/$1 "$TARGET_EXE_DIR"/plugins/$2
+    else
+        echo "restorePluginsQtInstallName: not enough parameters"
+    fi
 
-#This function sets correct relative pathes for linking qt libraries
-changeQtInstallNames () {
+    return 0;
+}
+
+# This function replaces @loader_path with @executable_path for UGENE plugins
+# Supposed that Qt is build as frameworks
+restorePluginsQtInstallNames () {
    if [ "$1" ]
    then
-        echo "Changing qt install names for $1"
+        # TODO: it is better to get Qt library names to change from otool output
 
-       install_name_tool -change $PATH_TO_QT/libQtCore.4.dylib  @executable_path/../Frameworks/libQtCore.4.dylib "$TARGET_EXE_DIR"/$1
-       install_name_tool -change $PATH_TO_QT/libQtScript.4.dylib  @executable_path/../Frameworks/libQtScript.4.dylib "$TARGET_EXE_DIR"/$1
-       install_name_tool -change $PATH_TO_QT/libQtXml.4.dylib  @executable_path/../Frameworks/libQtXml.4.dylib "$TARGET_EXE_DIR"/$1
-       install_name_tool -change $PATH_TO_QT/libQtXmlPatterns.4.dylib  @executable_path/../Frameworks/libQtXmlPatterns.4.dylib "$TARGET_EXE_DIR"/$1
-       install_name_tool -change $PATH_TO_QT/libQtGui.4.dylib  @executable_path/../Frameworks/libQtGui.4.dylib "$TARGET_EXE_DIR"/$1
-       install_name_tool -change $PATH_TO_QT/libQtNetwork.4.dylib  @executable_path/../Frameworks/libQtNetwork.4.dylib "$TARGET_EXE_DIR"/$1
-       install_name_tool -change $PATH_TO_QT/libQtWebKit.4.dylib  @executable_path/../Frameworks/libQtWebKit.4.dylib "$TARGET_EXE_DIR"/$1
-       install_name_tool -change $PATH_TO_QT/libphonon.4.dylib  @executable_path/../Frameworks/libphonon.4.dylib "$TARGET_EXE_DIR"/$1
-       install_name_tool -change $PATH_TO_QT/libQtTest.4.dylib  @executable_path/../Frameworks/libQtTest.4.dylib "$TARGET_EXE_DIR"/$1
-       install_name_tool -change $PATH_TO_QT/libQtOpenGL.4.dylib @executable_path/../Frameworks/libQtOpenGL.4.dylib "$TARGET_EXE_DIR"/$1
-       install_name_tool -change $PATH_TO_QT/libQtSvg.4.dylib  @executable_path/../Frameworks/libQtSvg.4.dylib "$TARGET_EXE_DIR"/$1
+        echo "Restore qt install names for plugin $1"
+        PLUGIN_LIB="lib$1.dylib"
+
+        restorePluginsQtInstallName QtCore $PLUGIN_LIB
+        restorePluginsQtInstallName QtGui $PLUGIN_LIB
+        restorePluginsQtInstallName QtNetwork $PLUGIN_LIB
+        restorePluginsQtInstallName QtOpenGL $PLUGIN_LIB
+        restorePluginsQtInstallName QtScript $PLUGIN_LIB
+        restorePluginsQtInstallName QtScriptTools $PLUGIN_LIB
+        restorePluginsQtInstallName QtSvg $PLUGIN_LIB
+        restorePluginsQtInstallName QtTest $PLUGIN_LIB
+        restorePluginsQtInstallName QtWebKit $PLUGIN_LIB
+        restorePluginsQtInstallName QtXml $PLUGIN_LIB
+        restorePluginsQtInstallName QtXmlPatterns $PLUGIN_LIB
 
    else
-       echo "changeQtInstallNames: no parameter passed."
+       echo "restorePluginsQtInstallNames: no parameter passed."
    fi
 
    return 0
 }
-
